+++ Makefile
@@ -7,7 +7,7 @@ deps:
 	go get github.com/golang/lint/golint
 
 test:
-	go test -race -cover -tags=test ./...
+	go test -race -cover ./...
 
 validate:
 	go vet ./...
