+++ client/container_create_test.go
@@ -8,14 +8,13 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/docker/engine-api/client/transport"
 	"github.com/docker/engine-api/types"
 	"github.com/docker/engine-api/types/container"
 )
 
 func TestContainerCreateError(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, transport.ErrorMock(http.StatusInternalServerError, "Server error")),
+		transport: newMockClient(nil, errorMock(http.StatusInternalServerError, "Server error")),
 	}
 	_, err := client.ContainerCreate(nil, nil, nil, "nothing")
 	if err == nil || err.Error() != "Error response from daemon: Server error" {
@@ -24,7 +23,7 @@ func TestContainerCreateError(t *testing.T) {
 
 	// 404 doesn't automagitally means an unknown image
 	client = &Client{
-		transport: transport.NewMockClient(nil, transport.ErrorMock(http.StatusNotFound, "Server error")),
+		transport: newMockClient(nil, errorMock(http.StatusNotFound, "Server error")),
 	}
 	_, err = client.ContainerCreate(nil, nil, nil, "nothing")
 	if err == nil || err.Error() != "Error response from daemon: Server error" {
@@ -34,7 +33,7 @@ func TestContainerCreateError(t *testing.T) {
 
 func TestContainerCreateImageNotFound(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, transport.ErrorMock(http.StatusNotFound, "No such image")),
+		transport: newMockClient(nil, errorMock(http.StatusNotFound, "No such image")),
 	}
 	_, err := client.ContainerCreate(&container.Config{Image: "unknown_image"}, nil, nil, "unknown")
 	if err == nil || !IsErrImageNotFound(err) {
@@ -44,7 +43,7 @@ func TestContainerCreateImageNotFound(t *testing.T) {
 
 func TestContainerCreateWithName(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, func(req *http.Request) (*http.Response, error) {
+		transport: newMockClient(nil, func(req *http.Request) (*http.Response, error) {
 			name := req.URL.Query().Get("name")
 			if name != "container_name" {
 				return nil, fmt.Errorf("container name not set in URL query properly. Expected `container_name`, got %s", name)
