+++ client/container_export_test.go
@@ -4,14 +4,12 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/docker/engine-api/client/transport"
-
 	"golang.org/x/net/context"
 )
 
 func TestContainerExportError(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, transport.ErrorMock(http.StatusInternalServerError, "Server error")),
+		transport: newMockClient(nil, errorMock(http.StatusInternalServerError, "Server error")),
 	}
 	_, err := client.ContainerExport(context.Background(), "nothing")
 	if err == nil || err.Error() != "Error response from daemon: Server error" {
