+++ client/container_kill_test.go
@@ -6,13 +6,11 @@ import (
 	"io/ioutil"
 	"net/http"
 	"testing"
-
-	"github.com/docker/engine-api/client/transport"
 )
 
 func TestContainerKillError(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, transport.ErrorMock(http.StatusInternalServerError, "Server error")),
+		transport: newMockClient(nil, errorMock(http.StatusInternalServerError, "Server error")),
 	}
 	err := client.ContainerKill("nothing", "SIGKILL")
 	if err == nil || err.Error() != "Error response from daemon: Server error" {
@@ -22,7 +20,7 @@ func TestContainerKillError(t *testing.T) {
 
 func TestContainerKill(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, func(req *http.Request) (*http.Response, error) {
+		transport: newMockClient(nil, func(req *http.Request) (*http.Response, error) {
 			signal := req.URL.Query().Get("signal")
 			if signal != "SIGKILL" {
 				return nil, fmt.Errorf("signal not set in URL query properly. Expected 'SIGKILL', got %s", signal)
