+++ client/container_pause_test.go
@@ -7,13 +7,11 @@ import (
 	"net/http"
 	"strings"
 	"testing"
-
-	"github.com/docker/engine-api/client/transport"
 )
 
 func TestContainerPauseError(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, transport.ErrorMock(http.StatusInternalServerError, "Server error")),
+		transport: newMockClient(nil, errorMock(http.StatusInternalServerError, "Server error")),
 	}
 	err := client.ContainerPause("nothing")
 	if err == nil || err.Error() != "Error response from daemon: Server error" {
@@ -24,7 +22,7 @@ func TestContainerPauseError(t *testing.T) {
 func TestContainerPause(t *testing.T) {
 	expectedURL := "/containers/container_id/pause"
 	client := &Client{
-		transport: transport.NewMockClient(nil, func(req *http.Request) (*http.Response, error) {
+		transport: newMockClient(nil, func(req *http.Request) (*http.Response, error) {
 			if !strings.HasPrefix(req.URL.Path, expectedURL) {
 				return nil, fmt.Errorf("Expected URL '%s', got '%s'", expectedURL, req.URL)
 			}
