+++ client/container_rename_test.go
@@ -7,13 +7,11 @@ import (
 	"net/http"
 	"strings"
 	"testing"
-
-	"github.com/docker/engine-api/client/transport"
 )
 
 func TestContainerRenameError(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, transport.ErrorMock(http.StatusInternalServerError, "Server error")),
+		transport: newMockClient(nil, errorMock(http.StatusInternalServerError, "Server error")),
 	}
 	err := client.ContainerRename("nothing", "newNothing")
 	if err == nil || err.Error() != "Error response from daemon: Server error" {
@@ -24,7 +22,7 @@ func TestContainerRenameError(t *testing.T) {
 func TestContainerRename(t *testing.T) {
 	expectedURL := "/containers/container_id/rename"
 	client := &Client{
-		transport: transport.NewMockClient(nil, func(req *http.Request) (*http.Response, error) {
+		transport: newMockClient(nil, func(req *http.Request) (*http.Response, error) {
 			if !strings.HasPrefix(req.URL.Path, expectedURL) {
 				return nil, fmt.Errorf("Expected URL '%s', got '%s'", expectedURL, req.URL)
 			}
