+++ client/image_build_test.go
@@ -11,7 +11,6 @@ import (
 
 	"golang.org/x/net/context"
 
-	"github.com/docker/engine-api/client/transport"
 	"github.com/docker/engine-api/types"
 	"github.com/docker/engine-api/types/container"
 	"github.com/docker/go-units"
@@ -19,7 +18,7 @@ import (
 
 func TestImageBuildError(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, transport.ErrorMock(http.StatusInternalServerError, "Server error")),
+		transport: newMockClient(nil, errorMock(http.StatusInternalServerError, "Server error")),
 	}
 	_, err := client.ImageBuild(context.Background(), types.ImageBuildOptions{})
 	if err == nil || err.Error() != "Error response from daemon: Server error" {
@@ -158,7 +157,7 @@ func TestImageBuild(t *testing.T) {
 	for _, buildCase := range buildCases {
 		expectedURL := "/build"
 		client := &Client{
-			transport: transport.NewMockClient(nil, func(r *http.Request) (*http.Response, error) {
+			transport: newMockClient(nil, func(r *http.Request) (*http.Response, error) {
 				if !strings.HasPrefix(r.URL.Path, expectedURL) {
 					return nil, fmt.Errorf("Expected URL '%s', got '%s'", expectedURL, r.URL)
 				}
