+++ client/image_create_test.go
@@ -10,13 +10,12 @@ import (
 
 	"golang.org/x/net/context"
 
-	"github.com/docker/engine-api/client/transport"
 	"github.com/docker/engine-api/types"
 )
 
 func TestImageCreateError(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, transport.ErrorMock(http.StatusInternalServerError, "Server error")),
+		transport: newMockClient(nil, errorMock(http.StatusInternalServerError, "Server error")),
 	}
 	_, err := client.ImageCreate(context.Background(), types.ImageCreateOptions{})
 	if err == nil || err.Error() != "Error response from daemon: Server error" {
@@ -30,7 +29,7 @@ func TestImageCreate(t *testing.T) {
 	expectedTag := "another:image"
 	expectedRegistryAuth := "eyJodHRwczovL2luZGV4LmRvY2tlci5pby92MS8iOnsiYXV0aCI6ImRHOTBid289IiwiZW1haWwiOiJqb2huQGRvZS5jb20ifX0="
 	client := &Client{
-		transport: transport.NewMockClient(nil, func(r *http.Request) (*http.Response, error) {
+		transport: newMockClient(nil, func(r *http.Request) (*http.Response, error) {
 			if !strings.HasPrefix(r.URL.Path, expectedURL) {
 				return nil, fmt.Errorf("Expected URL '%s', got '%s'", expectedURL, r.URL)
 			}
