+++ client/image_save_test.go
@@ -10,13 +10,12 @@ import (
 
 	"golang.org/x/net/context"
 
-	"github.com/docker/engine-api/client/transport"
 	"strings"
 )
 
 func TestImageSaveError(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, transport.ErrorMock(http.StatusInternalServerError, "Server error")),
+		transport: newMockClient(nil, errorMock(http.StatusInternalServerError, "Server error")),
 	}
 	_, err := client.ImageSave(context.Background(), []string{"nothing"})
 	if err == nil || err.Error() != "Error response from daemon: Server error" {
@@ -27,7 +26,7 @@ func TestImageSaveError(t *testing.T) {
 func TestImageSave(t *testing.T) {
 	expectedURL := "/images/get"
 	client := &Client{
-		transport: transport.NewMockClient(nil, func(r *http.Request) (*http.Response, error) {
+		transport: newMockClient(nil, func(r *http.Request) (*http.Response, error) {
 			if !strings.HasPrefix(r.URL.Path, expectedURL) {
 				return nil, fmt.Errorf("Expected URL '%s', got '%s'", expectedURL, r.URL)
 			}
