+++ client/network_inspect_test.go
@@ -9,13 +9,12 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/docker/engine-api/client/transport"
 	"github.com/docker/engine-api/types"
 )
 
 func TestNetworkInspectError(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, transport.ErrorMock(http.StatusInternalServerError, "Server error")),
+		transport: newMockClient(nil, errorMock(http.StatusInternalServerError, "Server error")),
 	}
 
 	_, err := client.NetworkInspect("nothing")
@@ -26,7 +25,7 @@ func TestNetworkInspectError(t *testing.T) {
 
 func TestNetworkInspectContainerNotFound(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, transport.ErrorMock(http.StatusNotFound, "Server error")),
+		transport: newMockClient(nil, errorMock(http.StatusNotFound, "Server error")),
 	}
 
 	_, err := client.NetworkInspect("unknown")
@@ -38,7 +37,7 @@ func TestNetworkInspectContainerNotFound(t *testing.T) {
 func TestNetworkInspect(t *testing.T) {
 	expectedURL := "/networks/network_id"
 	client := &Client{
-		transport: transport.NewMockClient(nil, func(req *http.Request) (*http.Response, error) {
+		transport: newMockClient(nil, func(req *http.Request) (*http.Response, error) {
 			if !strings.HasPrefix(req.URL.Path, expectedURL) {
 				return nil, fmt.Errorf("Expected URL '%s', got '%s'", expectedURL, req.URL)
 			}
