+++ client/volume_remove_test.go
@@ -7,13 +7,11 @@ import (
 	"net/http"
 	"strings"
 	"testing"
-
-	"github.com/docker/engine-api/client/transport"
 )
 
 func TestVolumeRemoveError(t *testing.T) {
 	client := &Client{
-		transport: transport.NewMockClient(nil, transport.ErrorMock(http.StatusInternalServerError, "Server error")),
+		transport: newMockClient(nil, errorMock(http.StatusInternalServerError, "Server error")),
 	}
 
 	err := client.VolumeRemove("volume_id")
@@ -26,7 +24,7 @@ func TestVolumeRemove(t *testing.T) {
 	expectedURL := "/volumes/volume_id"
 
 	client := &Client{
-		transport: transport.NewMockClient(nil, func(req *http.Request) (*http.Response, error) {
+		transport: newMockClient(nil, func(req *http.Request) (*http.Response, error) {
 			if !strings.HasPrefix(req.URL.Path, expectedURL) {
 				return nil, fmt.Errorf("Expected URL '%s', got '%s'", expectedURL, req.URL)
 			}
